--- consts/filterAction.go
@@ -1,10 +1,0 @@
-package consts
-
-type FilterAction int
-
-const (
-	FilterAdd FilterAction = iota
-	FilterConfirm
-	FilterDelete
-	FilterCancel
-)
