--- consts/modelName.go
@@ -1,8 +1,0 @@
-package consts
-
-type modelName int
-
-const (
-	UserModelName modelName = iota
-	OrganisationModelName
-)
