+++ filters/filterMsg.go
@@ -1,26 +1,31 @@
 package filters
 
-import (
-	"gh-hubbub/consts"
+type FilterAction int
+
+const (
+	FilterAdd FilterAction = iota
+	FilterConfirm
+	FilterDelete
+	FilterCancel
 )
 
 type FilterMsg struct {
-	Action consts.FilterAction
+	Action FilterAction
 	Filter Filter
 }
 
 func NewAddFilterMsg(filter Filter) FilterMsg {
-	return FilterMsg{Action: consts.FilterAdd, Filter: filter}
+	return FilterMsg{Action: FilterAdd, Filter: filter}
 }
 
 func NewConfirmFilterMsg(filter Filter) FilterMsg {
-	return FilterMsg{Action: consts.FilterConfirm, Filter: filter}
+	return FilterMsg{Action: FilterConfirm, Filter: filter}
 }
 
 func NewDeleteFilterMsg(filter Filter) FilterMsg {
-	return FilterMsg{Action: consts.FilterDelete, Filter: filter}
+	return FilterMsg{Action: FilterDelete, Filter: filter}
 }
 
 func NewCancelFilterMsg(filter Filter) FilterMsg {
-	return FilterMsg{Action: consts.FilterCancel, Filter: filter}
+	return FilterMsg{Action: FilterCancel, Filter: filter}
 }
