+++ models/filtersModel.go
@@ -1,0 +1,16 @@
+package models
+
+func (m *RepoModel) InitFilterEditor() {
+	// tab := m.repository.SettingsTabs[m.activeTab]
+	// index := m.settingsTable.Cursor()
+	// setting := tab.Settings[index]
+
+	// switch value := setting.Value.(type) {
+	// case bool:
+	// 	m.FilterModel = filters.NewBoolModel(tab.Name, setting.Name, value)
+	// case int:
+	// 	m.FilterModel = filters.NewIntModel(tab.Name, setting.Name, value, value)
+	// case time.Time:
+	// 	m.FilterModel = filters.NewDateModel(tab.Name, setting.Name, value, value)
+	// }
+}
