+++ models/repoModel.go
@@ -1,11 +1,8 @@
 package models
 
 import (
-	"gh-hubbub/consts"
-	"gh-hubbub/filters"
 	"gh-hubbub/structs"
 	"gh-hubbub/style"
-	"time"
 
 	"github.com/charmbracelet/bubbles/table"
 	tea "github.com/charmbracelet/bubbletea"
@@ -13,13 +10,10 @@ import (
 )
 
 type RepoModel struct {
-	repository  structs.RepositorySettings
-	FilterModel tea.Model
-
+	repository    structs.RepositorySettings
 	settingsTable table.Model
 
 	activeTab int
-	focus     consts.Focus
 	width     int
 	height    int
 }
@@ -44,10 +38,6 @@ func (m RepoModel) Init() tea.Cmd {
 	return nil
 }
 
-func (m *RepoModel) filterHasFocus() bool {
-	return m.focus == consts.FocusFilter
-}
-
 func (m *RepoModel) SelectRepo(repository structs.RepositorySettings) {
 	m.repository = repository
 	m.settingsTable = NewSettingsTable(m.repository.SettingsTabs[m.activeTab].Settings, m.width)
@@ -58,54 +48,18 @@ func (m *RepoModel) SelectTab(index int) {
 	m.settingsTable = NewSettingsTable(m.repository.SettingsTabs[m.activeTab].Settings, m.width)
 }
 
-func (m *RepoModel) InitFilterEditor() {
-	tab := m.repository.SettingsTabs[m.activeTab]
-	index := m.settingsTable.Cursor()
-	setting := tab.Settings[index]
-
-	switch value := setting.Value.(type) {
-	case bool:
-		m.FilterModel = filters.NewBoolModel(tab.Name, setting.Name, value)
-	case int:
-		m.FilterModel = filters.NewIntModel(tab.Name, setting.Name, value, value)
-	case time.Time:
-		m.FilterModel = filters.NewDateModel(tab.Name, setting.Name, value, value)
-	}
-}
-
 func (m RepoModel) Update(msg tea.Msg) (RepoModel, tea.Cmd) {
 	var cmd tea.Cmd
 
 	switch msg := msg.(type) {
 	case tea.KeyMsg:
 		switch msg.Type {
-		case tea.KeyEsc:
-			return m, m.FocusList
-		default:
-			if m.filterHasFocus() {
-				m.FilterModel, cmd = m.FilterModel.Update(msg)
-			} else {
-				m, cmd = m.UpdateRepoModel(msg.Type)
-			}
+		case tea.KeyRight:
+			m.SelectTab(min(m.activeTab+1, len(m.repository.SettingsTabs)-1))
+		case tea.KeyLeft:
+			m.SelectTab(max(m.activeTab-1, 0))
 		}
 	}
-	return m, cmd
-}
-
-func (m RepoModel) UpdateRepoModel(keyType tea.KeyType) (RepoModel, tea.Cmd) {
-	var cmd tea.Cmd
-
-	switch keyType {
-	case tea.KeyRight:
-		m.SelectTab(min(m.activeTab+1, len(m.repository.SettingsTabs)-1))
-	case tea.KeyLeft:
-		m.SelectTab(max(m.activeTab-1, 0))
-	case tea.KeyDown:
-		m.settingsTable.MoveDown(1)
-	case tea.KeyUp:
-		m.settingsTable.MoveUp(1)
-	}
-
 	return m, cmd
 }
 
@@ -115,20 +69,14 @@ func (m RepoModel) View() string {
 		return ""
 	}
 
-	frameWidth, frameHeight := style.Settings.GetFrameSize()
-	var tabs = RenderTabs(m.repository.SettingsTabs, m.width, m.activeTab)
+	// frameWidth, frameHeight := style.Settings.GetFrameSize()
 
-	// return style.Settings.Width(m.width - frameWidth).Height(m.height - frameHeight).Render(m.settingsTable.View())
+	settings := style.Settings.
+		Width(m.width).
+		// Height(5).
+		Render(m.settingsTable.View())
 
-	// if m.filterHasFocus() {
-	// 	filter := lipgloss.NewStyle().Width(m.width - 2).Height(20).Render(m.FilterModel.View())
-	// 	// filter := lipgloss.NewStyle().Width(m.width - 2).Height(m.height - 7).Render(m.FilterModel.View())
-	// 	return lipgloss.JoinVertical(lipgloss.Left, tabs, filter)
-	// } else {
-	// 	settings := style.Settings.Width(m.width - 2).Height(20).Render(m.settingsTable.View())
-	settings := style.Settings.Width(m.width - frameWidth).Height(m.height - frameHeight - 2).Render(m.settingsTable.View())
-	return lipgloss.JoinVertical(lipgloss.Left, tabs, settings)
-	// }
+	return lipgloss.JoinVertical(lipgloss.Left, settings)
 }
 
 func NewSettingsTable(activeSettings []structs.Setting, width int) table.Model {
@@ -146,16 +94,5 @@ func NewSettingsTable(activeSettings []structs.Setting, width int) table.Model {
 	table := table.New(table.WithColumns(columns), table.WithRows(rows),
 		table.WithFocused(true), table.WithStyles(table.DefaultStyles()))
 
-	table.SetHeight(10)
-
 	return table
-}
-
-func (m *RepoModel) SendFocusMsg() tea.Msg {
-	return NewFocusMsg(m.focus)
-}
-
-func (m *RepoModel) FocusList() tea.Msg {
-	m.focus = consts.FocusList
-	return m.SendFocusMsg()
 }
